website: blow away $JIRI_ROOT in setup instructions
Necessary because bootstrap.sh expects $JIRI_ROOT to not
already exist.

Our Jenkins test harness currently does not start each test
from a clean slate/workspace, so it's necessary to manually
blow away old unwanted state. (It would be better if the
default was a clean slate, with a whitelist for state that
should persist across tests.)

Change-Id: Ic67603bb99678d438adbfe0a411b9e4da0d79b1c