website: adds tests for inbound links
Test depends on building the site, scraping links, then visiting any internal
links to verify they are not broken.

An additional constraint for ensuring all href's resolve to an absolute path
was added to the link tests to make it easier to move content around later.

Additional, smaller changes:

* Makefile updated to properly define dependencies for build artifacts.
* Where possible .PHONY deps have been removed, ideally .PHONY targets are
  not depended on by build artifacts.

I intentionally didn't port the TOC tests from www. The changes and addition
of React casued enough complexity to warrent punting this to a new CL later.

Change-Id: Ic3013b9bbf21aa8f9b7ad9ebe0f95ee4fb5b1941