content: Add a page describing performance of the Vanadium framework.
Tons of micro-benchmarks are used to measure performance of different
parts of the Vanadium system. This page calls out a few specific ones
that we, as a team, should use when describing/measuring/comparing
performance at a high-level.

This newly added page is not linked into the main site just yet,
so you need to know its URL (https://vanadium.github.io/performance.html
once this change is deployed).

Change-Id: I54cf22489b7b96e97d7adee819ec43c11728f28f