website: Quick Start and Your First App
This CL drafts the Quick Start and Your First App with
semi-pseudo code that needs updating and testing when high-level
API is done. App animations also needs to be updated when
we have the working app.

This CL also adds the ability to have parts of the code be "dimmed"
so we can highlight important aspects better.

https://screenshot.googleplex.com/SGGHnxTiYbP.png
https://screencast.googleplex.com/cast/6504361354067968

Change-Id: Id2dea6eb1f72b3319cd58c66ac7bd2c79e151d87