Fix installation instructions for installing profile packages.
The step-by-step install instructions tell the user to run `jiri profile
os-packages v23:base`, and then run the output as sudo.  This works for
the step-by-step instructions, but not for the main installation script,
which is concatenation of the step-by-steps.

This CL adds an explicit line "sudo $(jiri profile os-packages
v23:base)" to the step-by-step instructions that will install the
package prereqs, fixing the main installation script.

Fixes https://github.com/vanadium/issues/issues/1346

Change-Id: I32188e92a6e15bf8b28f38a8f0062609a896488c