website: roadmap
This is a refactor of information that was already on the website.
References to future and in-progress work have been moved from
inline documentation to a separate roadmap doc.
Addied info on the current state of blob implementation which
is available in Go, but has no high-level APIs in Swift and Java.

Change-Id: I7653af33243a85c0d0c3c5426ef057fb45c0c110